refactoring Nav
